Update to actually retry the requests as many times the input parameter says. Better error reporting.
Summary:
The retry POST/GET was in reality never used and the error status from the
was never reported. With this fix we can retry the urls and see the status in
uservice to find/track odin-api problems

Test Plan: Manual test of the code done in matlab with a echo-server.

Reviewers: fredrik, bengt.rydberg, skymandr

Reviewed By: skymandr

Differential Revision: https://phabricator.molflow.com/D494