absorption lookup table and q setting structure fix in image building
Summary:
absorption lookup table is now stored under directory QsmrData in the docker image where
it is generated (same as in the worker image used for level2 processing).
The Q setting structure used for generating the absorption lookup table
is saved as a mat file in this process, and is also used for level2 processing.
It was a mismatch in the field FOLDER_ABSLOOKUP which is now corrceted.

Test Plan:
I have run the script create_qsmr_project.sh locally on my computer and verified that
the field Q.FOLDER_ABSLOOKUP gets correct name.

Reviewers: joakim

Reviewed By: joakim

Differential Revision: https://phabricator.molflow.com/D1233